Fix type mismatches and clamping issues in colony system calculations
- Converted integer literals to f64 for compatibility in clamping operations
- Updated `.clamp()` bounds to use `0.0` and `100.0` for f64 consistency
- Resolved Git remote configuration to point to correct repository URL

+++ src/systems/cell.rs
@@ -48,12 +48,12 @@ impl Cell {
             neighbors: Vec::new(),
             energy: 100.0,
             dimensional_position: DimensionalPosition {
-                emergence: 50,
-                coherence: 50,
-                resilience: 50,
-                intelligence: 50,
-                efficiency: 50,
-                integration: 50,
+                emergence: 50.0,
+                coherence: 50.0,
+                resilience: 50.0,
+                intelligence: 50.0,
+                efficiency: 50.0,
+                integration: 50.0,
             },
             dopamine: 0.5,
             enhanced_state: EnhancedCellState::new(),
@@ -530,10 +530,10 @@ impl Cell {
         let efficiency_adj = (dimension_adjustment * 10.0) as i32;
         
         // Apply adjustments and clamp to 0-100 range
-        self.dimensional_position.emergence = (self.dimensional_position.emergence + emergence_adj).clamp(0, 100);
-        self.dimensional_position.coherence = (self.dimensional_position.coherence + coherence_adj).clamp(0, 100);
-        self.dimensional_position.intelligence = (self.dimensional_position.intelligence + intelligence_adj).clamp(0, 100);
-        self.dimensional_position.efficiency = (self.dimensional_position.efficiency + efficiency_adj).clamp(0, 100);
+        self.dimensional_position.emergence = (self.dimensional_position.emergence + emergence_adj as f64).clamp(0.0, 100.0);
+        self.dimensional_position.coherence = (self.dimensional_position.coherence + coherence_adj as f64).clamp(0.0, 100.0);
+        self.dimensional_position.intelligence = (self.dimensional_position.intelligence + intelligence_adj as f64).clamp(0.0, 100.0);
+        self.dimensional_position.efficiency = (self.dimensional_position.efficiency + efficiency_adj as f64).clamp(0.0, 100.0);
         
         // Adjust dopamine based on context alignment
         self.dopamine = self.dopamine * 0.9 + self.context_alignment_score * 0.1;
